Add support for attribute exchange as a way to query user details.
If the endpoint claims to support attribute exchange, use that extension 
to request the details instead of simple registration.

We use the standard axschema.org attribute names, and ask for the user's 
full name in both complete and split first/last name versions.  This 
should let us support both the Google and Yahoo provider 
implementations.